requirements.txt
# ============================================================================
# CA FIRM BACKEND - REQUIREMENTS
# Python 3.10+
# Last Updated: 2025
# ============================================================================

# ============================================================================
# CORE DJANGO (REQUIRED)
# ============================================================================
Django~=5.1.0
djangorestframework~=3.15.0
djangorestframework-simplejwt~=5.3.1
python-dotenv~=1.0.1

# ============================================================================
# DATABASE - PostgreSQL (REQUIRED)
# ============================================================================
psycopg2-binary~=2.9.9
dj-database-url~=2.2.0

# ============================================================================
# REDIS & CACHING (REQUIRED for production)
# ============================================================================
redis~=5.0.8
django-redis~=5.4.0
hiredis~=2.3.2  # C parser for better performance

# ============================================================================
# CELERY - Background Tasks (REQUIRED)
# ============================================================================
celery~=5.4.0
django-celery-beat~=2.7.0
django-celery-results~=2.5.1
kombu~=5.3.0  # Messaging library
flower # Celery monitoring

# ============================================================================
# SECURITY & AUTHENTICATION (REQUIRED)
# ============================================================================
django-cors-headers~=4.5.0
argon2-cffi~=23.1.0  # Better password hashing
PyJWT~=2.8.0  # JWT handling

# ============================================================================
# RATE LIMITING & PROTECTION (RECOMMENDED)
# ============================================================================
django-ratelimit~=4.1.0
django-defender~=0.9.7  # Brute force protection

# ============================================================================
# FILE HANDLING (REQUIRED)
# ============================================================================
django-storages[boto3]~=1.14.4
Pillow~=10.4.0
python-magic~=0.4.27

# ============================================================================
# CA FIRM SPECIFIC (REQUIRED)
# ============================================================================
reportlab~=4.2.0  # PDF generation
openpyxl~=3.1.5  # Excel handling
pandas~=2.2.0  # Data analysis
PyPDF2~=3.0.1  # PDF processing
python-dateutil~=2.8.2  # Date utilities

# ============================================================================
# API DOCUMENTATION (REQUIRED)
# ============================================================================
drf-spectacular~=0.27.0
pyyaml~=6.0.1  # YAML support for OpenAPI

# ============================================================================
# PRODUCTION SERVER (REQUIRED)
# ============================================================================
gunicorn~=23.0.0
whitenoise~=6.7.0
gevent~=24.2.1  # Async worker support

# ============================================================================
# MONITORING & ERROR TRACKING (RECOMMENDED)
# ============================================================================
sentry-sdk~=2.14.0

# ============================================================================
# UTILITIES (RECOMMENDED)
# ============================================================================
python-slugify~=8.0.1  # URL-safe strings
pytz~=2024.1  # Timezone handling
requests~=2.31.0  # HTTP client

# ============================================================================
# DEVELOPMENT TOOLS (OPTIONAL - Comment out in production)
# ============================================================================
# django-debug-toolbar~=4.4.0
# django-extensions~=3.2.3
# ipython~=8.26.0

# ============================================================================
# TESTING (OPTIONAL - For CI/CD)
# ============================================================================
# pytest~=8.3.0
# pytest-django~=4.9.0
# pytest-cov~=5.0.0
# pytest-xdist~=3.6.0  # Parallel testing
# factory-boy~=3.3.1
# faker~=30.0.0
# coverage~=7.6.0

# ============================================================================
# CODE QUALITY (OPTIONAL - For development)
# ============================================================================
# black~=24.8.0  # Code formatter
# flake8~=7.1.0  # Linter
# isort~=5.13.0  # Import sorter
# mypy~=1.11.0  # Type checker
# pylint~=3.2.0  # Code analysis

# ============================================================================
# NOTES
# ============================================================================
# 1. Pin exact versions in production: Django==5.1.0
# 2. Use requirements-dev.txt for development-only packages
# 3. Regularly update with: pip list --outdated
# 4. Security audit with: pip-audit
# 5. For ARM64 (Apple Silicon): ensure compatible wheels